+++ packages/amplify-cli/src/__tests__/commands/upgrade.pkg.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-interpolation-in-snapshots */
 import * as fs from 'fs-extra';
 import fetch, { Response } from 'node-fetch';
 import { $TSContext } from '@aws-amplify/amplify-cli-core';
@@ -156,10 +155,10 @@ describe('run upgrade using packaged CLI', () => {
 
     let movedBinToTemp = false;
     fsMock.move
-      .mockImplementationOnce(async () => {
+      .mockImplementationOnce(() => {
         movedBinToTemp = true;
       })
-      .mockImplementationOnce(async () => {
+      .mockImplementationOnce(() => {
         if (!movedBinToTemp) throw new Error('fs.move was not called before copying extracted file to bin location');
       });