+++ packages/amplify-e2e-tests/src/__tests__/git-clone-attach.test.ts
@@ -91,7 +91,7 @@ describe('attach amplify to git-cloned project', () => {
     const preCleanTpi = getTeamProviderInfo(projRoot);
     const importBucketRegion = (Object.values(preCleanTpi[envName].categories.storage)[0] as any).region;
     const appId = preCleanTpi[envName].awscloudformation.AmplifyAppId;
-    gitCleanFdx(projRoot);
+    await gitCleanFdx(projRoot);
 
     const socialProviders = getSocialProviders();
     const categoriesConfig = {