+++ packages/amplify-e2e-tests/src/__tests__/storage-1b.test.ts
@@ -21,6 +21,7 @@ function getServiceMeta(projectRoot: string, category: string, service: string):
       return meta.storage[storageResourceName];
     }
   }
+  return undefined;
 }
 
 describe('amplify add/update storage(S3)', () => {