+++ packages/amplify-e2e-tests/src/__tests__/storage-simulator/S3server.test.ts
@@ -2,10 +2,10 @@ import { AmplifyStorageSimulator } from 'amplify-storage-simulator';
 import * as AWS from 'aws-sdk';
 import * as fs from 'fs-extra';
 
-let port = 20005; // for testing
-let route = '/mock-testing';
-let bucket = 'mock-testing';
-let localDirS3 = __dirname + '/test-data/';
+const port = 20005; // for testing
+const route = '/mock-testing';
+const bucket = 'mock-testing';
+const localDirS3 = __dirname + '/test-data/';
 const actual_file = __dirname + '/test-data/2.png';
 
 let s3client;
@@ -20,7 +20,7 @@ beforeAll(async () => {
     region: 'eu-west-2',
   });
 
-  let ep = new AWS.Endpoint('http://localhost:20005');
+  const ep = new AWS.Endpoint('http://localhost:20005');
   s3client = new AWS.S3({
     apiVersion: '2006-03-01',
     endpoint: ep.href,
@@ -90,7 +90,7 @@ describe('Test list api', () => {
   });
 
   test('list object pagination', async () => {
-    let maxKeys = 2;
+    const maxKeys = 2;
     let total = 7;
     let response = await s3client
       .listObjects({