+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@4.3.2...amplify-e2e-tests@4.3.3) (2023-07-28)
+
+
+### Bug Fixes
+
+* surface override build error if scripting features are disabled ([#13013](https://github.com/aws-amplify/amplify-cli/issues/13013)) ([#13023](https://github.com/aws-amplify/amplify-cli/issues/13023)) ([9e99800](https://github.com/aws-amplify/amplify-cli/commit/9e998008c6aef13ee50e6ce88cc78d8e9367b66a))
+
+
+
+
+
+## [4.3.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@4.3.1...amplify-e2e-tests@4.3.2) (2023-07-27)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
 ## [4.3.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@4.3.0...amplify-e2e-tests@4.3.1) (2023-07-24)
 
 **Note:** Version bump only for package amplify-e2e-tests