+++ packages/amplify-environment-parameters/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-environment-parameters@1.7.2...@aws-amplify/amplify-environment-parameters@1.7.3) (2023-07-28)
+
+**Note:** Version bump only for package @aws-amplify/amplify-environment-parameters
+
+
+
+
+
+## [1.7.2](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-environment-parameters@1.7.1...@aws-amplify/amplify-environment-parameters@1.7.2) (2023-07-27)
+
+**Note:** Version bump only for package @aws-amplify/amplify-environment-parameters
+
+
+
+
+
 ## [1.7.1](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-environment-parameters@1.7.0...@aws-amplify/amplify-environment-parameters@1.7.1) (2023-07-24)
 
 **Note:** Version bump only for package @aws-amplify/amplify-environment-parameters