+++ packages/amplify-util-uibuilder/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.3](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.9.2...@aws-amplify/amplify-util-uibuilder@1.9.3) (2023-07-28)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder
+
+
+
+
+
+## [1.9.2](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.9.1...@aws-amplify/amplify-util-uibuilder@1.9.2) (2023-07-27)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder
+
+
+
+
+
 ## [1.9.1](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.9.0...@aws-amplify/amplify-util-uibuilder@1.9.1) (2023-07-24)
 
 **Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder